fixed HA bugs and docs

+++ lib/fantom/lib/assets/download-snapshot.sh
@@ -7,13 +7,15 @@ echo "Downloading FANTOM snapshot."
 
 cd /data
 
-for FILE in `curl https://snapshot.fantom.network/files/snapsync/latest/listtgzfiles.txt`; do
-        echo $FILE;
-        axel -n 20 https://snapshot.fantom.network/files/snapsync/$FILE && \
-        tar --use-compress-program="pigz -d" -xvf ${FILE##*/} && \
-        rm ${FILE##*/} || \
-        echo "Problem with downloading or expanding file $FILE"
-done
+if [[ -n ${FANTOM_SNAPSHOTS_URI} && ${FANTOM_SNAPSHOTS_URI} != "none" ]]; then
+        for FILE in `curl ${FANTOM_SNAPSHOTS_URI}`; do
+                echo $FILE;
+                axel -n 20 ${FANTOM_SNAPSHOTS_URI%%/latest/listtgzfiles.txt}/$FILE && \
+                tar --use-compress-program="pigz -d" -xvf ${FILE##*/} && \
+                rm ${FILE##*/} || \
+                echo "Problem with downloading or expanding file $FILE"
+        done
+fi
 
 echo "Downloading FANTOM snapshot finished"
